+++ packages/storage/src/index.ts
@@ -2,3 +2,4 @@ export * from './prisma/index.js';
 export * from './minio/index.js';
 export * from './mime/index.js';
 export * from './mail-rules.js';
+export * from './transport-rules.js';